wear: hopefully fixed offset error in the vibration times
There was an error that accumulated over time, because you would not start the vibration after exactly 1000ms, but whenever the current time was OVER 1000ms. That could be e.g. at 1020ms or so. So in order to not introduce a delay that accumulates (because the next one would be e.g. at 2040ms), we just keep the constant offset (and start the next vibration at 2020ms).