mobile: changed MA window sizes
Assuming the average gesture is about 1500ms long and the wearable sends an update every 40-50ms (which are both reasonable assumptions). A window of 43 would mean that around 43*45ms = 1935ms would be incorporated in the MA. That means that the previous
gesture will leak into the current one.
I re-ran the hyperparameter optimization with bounds for the parameters ((1, 10), (1, 10), (1, 25)) and got the new values.

23*45ms = 1035ms is a lot less leakage.

Current maximum @ 0.683815153411
Parameters: {'features__ma2__timesteps': 3, 'features__ma1__timesteps': 2, 'features__ma3__timesteps': 23}